Add Customer Service section to tenancy levels
