+++ src/main/pages/features/TenancyLevels.js
@@ -4,13 +4,16 @@ import TenancyInfo from "./TenancyInfo";
 import ImageIcon from "../../components/ImageIcon";
 import {TabbedInfoBox} from "../../components/Misc";
 import {PageContainer} from "../../MainApp";
-import {FormatCurrency} from "../../utils/Utils";
 import FAQs from "./FAQs";
 import {observer} from "mobx-react";
 import {mainStore} from "../../stores/Main";
 
-import {PlayCircleIcon} from "../../static/icons/Icons";
+import {
+  PlayCircleIcon,
+} from "../../static/icons/Icons";
 import TechnologyIcons from "../../static/icons/technology/TechnologyIcons";
+import SupportGrid from "./SupportGrid";
+import {CustomerServiceSection} from "./Support";
 
 const BannerBox = ({title, icon, paragraph, link}) => {
   const icons = {
@@ -36,11 +39,29 @@ const BannerBox = ({title, icon, paragraph, link}) => {
   );
 };
 
+const TableSection = observer(() => {
+  return (
+    <div className="tenancies-comparison curved-box info-box light">
+      <div className="tenancies-comparison__table-container">
+        <div className="tenancies-comparison__header">Feature Comparison at a Glance</div>
+        <SupportGrid />
+      </div>
+
+      <div className="tenancies-comparison__table-container">
+        <div className="tenancies-comparison__header">Customer Service and Support</div>
+        <SupportGrid items={mainStore.l10n.features_support.gridItems} compactRows={true} />
+        <CustomerServiceSection smallFont={true} />
+      </div>
+    </div>
+  );
+});
+
 const TenanciesList = ({monthly=false}) => {
   return (
     <div className="tenancies-list">
       {mainStore.l10n.features.tenancies.levels.map(({header, sub_header, monthly_price, annual_price, added_benefit_text, added_benefit_percentage, features}) => (
         <TenancyInfo
+          key={header}
           header={header}
           subHeader={sub_header}
           monthlyPrice={monthly ? monthly_price : annual_price}
@@ -101,6 +122,7 @@ const TenancyLevels = observer(() => {
           <div className="features-connect">
             <Button className="light primary features-connect__button" to="/about/contact">Connect with us</Button>
           </div>
+          <TableSection />
         </div>
       </div>
     </PageContainer>
